refactor: VatWorkerManager -> VatWorkerService
