+++ packages/nodejs/src/index.ts
@@ -1,2 +1,2 @@
-export { NodejsVatWorkerManager } from './kernel/VatWorkerManager.ts';
+export { NodejsVatWorkerService } from './kernel/VatWorkerService.ts';
 export { makeKernel } from './kernel/make-kernel.ts';
