+++ packages/nodejs/src/kernel/VatWorkerService.test.ts
@@ -2,41 +2,56 @@ import '@metamask/kernel-shims/endoify';
 
 import { makeCounter } from '@metamask/kernel-utils';
 import type { VatId } from '@metamask/ocap-kernel';
-import { NodeWorkerDuplexStream } from '@metamask/streams';
 import { describe, expect, it, vi } from 'vitest';
 
-import { NodejsVatWorkerManager } from '../../src/kernel/VatWorkerManager.ts';
-import { getTestWorkerFile } from '../get-test-worker.ts';
+import { NodejsVatWorkerService } from './VatWorkerService.ts';
 
-describe('NodejsVatWorkerManager', () => {
-  const testWorkerFile = getTestWorkerFile('stream-sync');
+const mocks = vi.hoisted(() => ({
+  worker: {
+    once: (_: string, callback: () => unknown) => {
+      callback();
+    },
+    terminate: vi.fn(async () => undefined),
+  },
+  stream: {
+    synchronize: vi.fn(async () => undefined).mockResolvedValue(undefined),
+    return: vi.fn(async () => ({})),
+  },
+}));
+
+vi.mock('@metamask/streams', () => ({
+  NodeWorkerDuplexStream: vi.fn(() => mocks.stream),
+}));
+
+vi.mock('node:worker_threads', () => ({
+  Worker: vi.fn(() => mocks.worker),
+}));
+
+describe('NodejsVatWorkerService', () => {
+  it('constructs an instance without any arguments', () => {
+    const instance = new NodejsVatWorkerService({});
+    expect(instance).toBeInstanceOf(NodejsVatWorkerService);
+  });
+
+  const workerFilePath = 'unused';
   const vatIdCounter = makeCounter();
   const getTestVatId = (): VatId => `v${vatIdCounter()}`;
 
   describe('launch', () => {
     it('creates a NodeWorker and returns a NodeWorkerDuplexStream', async () => {
-      const service = new NodejsVatWorkerManager({
-        workerFilePath: testWorkerFile,
+      const service = new NodejsVatWorkerService({
+        workerFilePath,
       });
       const testVatId: VatId = getTestVatId();
       const stream = await service.launch(testVatId);
 
-      expect(stream).toBeInstanceOf(NodeWorkerDuplexStream);
+      expect(stream).toStrictEqual(mocks.stream);
     });
 
     it('rejects if synchronize fails', async () => {
       const rejected = 'test-reject-value';
-
-      vi.doMock('@metamask/streams', () => ({
-        NodeWorkerDuplexStream: vi.fn().mockImplementation(() => ({
-          synchronize: vi.fn(() => 'no').mockRejectedValue(rejected),
-        })),
-      }));
-      vi.resetModules();
-      const NVWS = (await import('../../src/kernel/VatWorkerManager.ts'))
-        .NodejsVatWorkerManager;
-
-      const service = new NVWS({ workerFilePath: testWorkerFile });
+      mocks.stream.synchronize.mockRejectedValue(rejected);
+      const service = new NodejsVatWorkerService({ workerFilePath });
       const testVatId: VatId = getTestVatId();
       await expect(async () => await service.launch(testVatId)).rejects.toThrow(
         rejected,
@@ -46,8 +61,8 @@ describe('NodejsVatWorkerManager', () => {
 
   describe('terminate', () => {
     it('terminates the target vat', async () => {
-      const service = new NodejsVatWorkerManager({
-        workerFilePath: testWorkerFile,
+      const service = new NodejsVatWorkerService({
+        workerFilePath,
       });
       const testVatId: VatId = getTestVatId();
 
@@ -59,8 +74,8 @@ describe('NodejsVatWorkerManager', () => {
     });
 
     it('throws when terminating an unknown vat', async () => {
-      const service = new NodejsVatWorkerManager({
-        workerFilePath: testWorkerFile,
+      const service = new NodejsVatWorkerService({
+        workerFilePath,
       });
       const testVatId: VatId = getTestVatId();
 
@@ -72,8 +87,8 @@ describe('NodejsVatWorkerManager', () => {
 
   describe('terminateAll', () => {
     it('terminates all vats', async () => {
-      const service = new NodejsVatWorkerManager({
-        workerFilePath: testWorkerFile,
+      const service = new NodejsVatWorkerService({
+        workerFilePath,
       });
       const vatIds: VatId[] = [getTestVatId(), getTestVatId(), getTestVatId()];
