+++ packages/ocap-kernel/src/Kernel.test.ts
@@ -11,7 +11,7 @@ import { Kernel } from './Kernel.ts';
 import type {
   VatId,
   VatConfig,
-  VatWorkerManager,
+  VatWorkerService,
   ClusterConfig,
 } from './types.ts';
 import { VatHandle } from './VatHandle.ts';
@@ -54,7 +54,7 @@ const makeMockClusterConfig = (): ClusterConfig => ({
 
 describe('Kernel', () => {
   let mockStream: DuplexStream<JsonRpcRequest, JsonRpcResponse>;
-  let mockWorkerService: VatWorkerManager;
+  let mockWorkerService: VatWorkerService;
   let launchWorkerMock: MockInstance;
   let terminateWorkerMock: MockInstance;
   let makeVatHandleMock: MockInstance;
@@ -72,7 +72,7 @@ describe('Kernel', () => {
         ({}) as unknown as DuplexStream<JsonRpcMessage, JsonRpcMessage>,
       terminate: async () => undefined,
       terminateAll: async () => undefined,
-    } as unknown as VatWorkerManager;
+    } as unknown as VatWorkerService;
 
     launchWorkerMock = vi
       .spyOn(mockWorkerService, 'launch')
