+++ packages/ocap-kernel/src/Kernel.ts
@@ -25,7 +25,7 @@ import type { KernelStore } from './store/index.ts';
 import type {
   VatId,
   KRef,
-  VatWorkerManager,
+  VatWorkerService,
   ClusterConfig,
   VatConfig,
   KernelStatus,
@@ -44,7 +44,7 @@ export class Kernel {
   readonly #vats: Map<VatId, VatHandle>;
 
   /** Service to spawn workers (in iframes) for vats to run in */
-  readonly #vatWorkerService: VatWorkerManager;
+  readonly #vatWorkerService: VatWorkerService;
 
   /** Storage holding the kernel's own persistent state */
   readonly #kernelStore: KernelStore;
@@ -74,7 +74,7 @@ export class Kernel {
   // eslint-disable-next-line no-restricted-syntax
   private constructor(
     commandStream: DuplexStream<JsonRpcCall, JsonRpcResponse>,
-    vatWorkerService: VatWorkerManager,
+    vatWorkerService: VatWorkerService,
     kernelDatabase: KernelDatabase,
     options: {
       resetStorage?: boolean;
@@ -113,7 +113,7 @@ export class Kernel {
    */
   static async make(
     commandStream: DuplexStream<JsonRpcCall, JsonRpcResponse>,
-    vatWorkerService: VatWorkerManager,
+    vatWorkerService: VatWorkerService,
     kernelDatabase: KernelDatabase,
     options: {
       resetStorage?: boolean;
